requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
pypdf
faiss-cpu
langsmith
langchain-pinecone
requests
black
openai
google-search-results
python-dotenv
pinecone
beautifulsoup4
altair<5
streamlit
streamlit_chat
firecrawl-py